Renamed math.hpp to constants.hpp
