+++ src/domain/angle.hpp
@@ -1,7 +1,7 @@
 #ifndef DOMAIN_DEGREES_HPP
 #define DOMAIN_DEGREES_HPP
 
-#include "domain/math.hpp"
+#include "domain/constants.hpp"
 
 #include <compare>
 #include <string>
